requirements.txt
attrs==22.2.0
exceptiongroup==1.1.0
iniconfig==2.0.0
packaging==23.0
pluggy==1.0.0
polars==0.16.4
pytest==7.2.1
PyYAML==6.0
tomli==2.0.1
typing_extensions==4.4.0
